attendence/requirements.txt
alabaster==0.7.12
argh==0.26.2
astor==0.8.1
astroid @ file:///C:/ci/astroid_1613501147567/work
async-generator==1.10
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1615918855173/work
Babel @ file:///tmp/build/80754af9/babel_1607110387436/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bcrypt @ file:///C:/ci/bcrypt_1597936262193/work
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247269562/work
Chameleon==3.6.2
chardet @ file:///C:/ci/chardet_1607706912142/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
cligj==0.5.0
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.9.2
conda-build==3.20.5
conda-package-handling @ file:///C:/ci/conda-package-handling_1603018126209/work
configobj==5.0.6
cryptography @ file:///C:/ci/cryptography_1615532454813/work
DateTime==4.3
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
diff-match-patch @ file:///tmp/build/80754af9/diff-match-patch_1594828741838/work
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
flake8 @ file:///tmp/build/80754af9/flake8_1615834841867/work
future==0.18.2
gast==0.2.2
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-pasta==0.2.0
grpcio==1.30.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize==1.2.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593431910144/work
intervaltree @ file:///tmp/build/80754af9/intervaltree_1598376443606/work
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort @ file:///tmp/build/80754af9/isort_1612243941122/work
jdcal==1.4.1
jedi @ file:///C:/ci/jedi_1592841982010/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
json-merge-patch==0.2
jsondate==0.1.2
jsondiff==1.2.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213536848/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras==2.2.5
keyring @ file:///C:/ci/keyring_1614617025616/work
lazy-object-proxy @ file:///C:/ci/lazy-object-proxy_1606850127301/work
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.0
mkl-service==2.3.0
MouseInfo==0.1.3
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601896775581/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
numpydoc @ file:///tmp/build/80754af9/numpydoc_1605117425582/work
olefile==0.46
opencv-python==4.3.0.36
openpyxl==3.0.4
opt-einsum==3.2.1
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandocfilters @ file:///C:/ci/pandocfilters_1605120535071/work
paramiko @ file:///tmp/build/80754af9/paramiko_1598886428689/work
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1594299032796/work
pkginfo==1.7.0
pluggy @ file:///C:/ci/pluggy_1615976526223/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.12.2
psutil @ file:///C:/ci/psutil_1612298125479/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
PyAutoGUI==0.9.50
pycodestyle @ file:///home/ktietz/src/ci_mi/pycodestyle_1612807597675/work
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1598885001695/work
pyflakes @ file:///home/ktietz/src/ci_ipy2/pyflakes_1612551159640/work
PyGetWindow==0.0.8
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pylint @ file:///C:/ci/pylint_1614971403575/work
PyMsgBox==1.0.8
PyNaCl @ file:///C:/ci/pynacl_1595009245871/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
PyScreeze==0.1.26
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1600278539111/work
python-language-server @ file:///tmp/build/80754af9/python-language-server_1600454544709/work
PyTweening==1.0.3
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtAwesome @ file:///tmp/build/80754af9/qtawesome_1615991616277/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1612458529756/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
rope @ file:///tmp/build/80754af9/rope_1602264064449/work
Rtree==0.9.4
ruamel-yaml==0.15.87
scipy==1.5.1
six @ file:///C:/ci/six_1605187303045/work
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1611258885636/work
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1606865132123/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1614023247537/work
Sphinx @ file:///tmp/build/80754af9/sphinx_1615238060936/work
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///home/ktietz/src/ci/sphinxcontrib-htmlhelp_1611920974801/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///home/ktietz/src/ci/sphinxcontrib-serializinghtml_1611920755253/work
spyder @ file:///C:/ci/spyder_1594825773579/work
spyder-kernels @ file:///D:/bld/spyder-kernels_1594392089024/work
tensorflow==1.10.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tornado @ file:///C:/ci/tornado_1606942379977/work
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
traitlets==4.3.3
typed-ast @ file:///C:/ci/typed-ast_1610466539988/work
ujson @ file:///C:/ci/ujson_1611259572767/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
watchdog @ file:///C:/ci/watchdog_1612471247473/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter @ file:///tmp/build/80754af9/xlsxwriter_1602692860603/work
yapf @ file:///tmp/build/80754af9/yapf_1615749224965/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
zope.interface==5.1.0

numpy~=1.17.4
requests~=2.25.1
pillow~=8.2.0
six~=1.15.0